fix: cast f32 accumulator output back to q_dtype before writing to output buffer
The f32 out_dtype causes the flash attention output to be f32 while
the output HBM buffer (bo_x2_ref) is sized for bf16. The bitcast
reshape fails with shape mismatch. Fix: accumulate in f32 for
precision, but cast final result back to q_dtype (bf16) for output.

+++ python/sgl_jax/srt/kernels/ragged_paged_attention/ragged_paged_attention_v3.py
@@ -1064,7 +1064,7 @@ def _ragged_paged_attention_kernel_loop(
                 acc * pl.reciprocal(l, approx=True)
                 if (l.dtype == jnp.float32 and out_dtype != jnp.float32)
                 else lax.div(acc, l)
-            ).astype(out_dtype)
+            ).astype(q_dtype)
 
             # Wait for previous bo to be fully sent before storing new bo.
             bo_sem_idx = sem_ids_ref[2]
